Refactor trustee migration to follow CAMS architecture
Move district-division-mapper into migrate-trustees module to consolidate
related functionality. Remove division enrichment from ATS domain, keeping
it in use-case layer. Make district-to-division mapping required throughout
call stack. Use case now fetches office data via factory pattern.

Jira ticket: CAMS-740
